admiralophtha: ADaM in R Asset Library - Ophthalmology

Aids the programming of Clinical Data Standards Interchange Consortium (CDISC) compliant Ophthalmology Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package).

See also

Author

Maintainer: Edoardo Mancini edoardo.mancini@roche.com (ORCID)

Authors:

Edoardo Mancini edoardo.mancini@roche.com (ORCID)

Yuki Matsunaga yuki.matsunaga@novartis.com

Other contributors:

F. Hoffmann-La Roche AG [copyright holder, funder]

Novartis AG [copyright holder, funder]