admiralophtha: ADaM in R Asset Library - Ophthalmology

Aids the programming of Clinical Data Standards Interchange Consortium (CDISC) compliant Ophthalmology Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package).

See also

Author

Maintainer: Edoardo Mancini edoardo.mancini@roche.com (ORCID)

Authors:

Ritika Aggarwal ritika.aggarwal@novartis.com

Jane Gao jane.gao@roche.com

William Holmes william.holmes@roche.com

Josie Jackson josie.jackson@roche.com

Sonali Jain sonali.jain@novartis.com

Yuki Matsunaga yuki.matsunaga@novartis.com

Gordon Miller gordon.miller@roche.com

Rachel Linacre rachel.linacre@roche.com

Lucy Palmen lucy.palmen@roche.com

Nandini R Thampi nandini.rajendhiran@novartis.com

Aldrich Salva salva.aldrich@gene.com