Add CRITx and CRITxFL pair to ADBCVA dataset

Helper function for derive_var_bcvacritxfl that adds a criterion variable CRITx and its corresponding flag CRITxFL to an ADBCVA dataset.

Usage

derive_var_bcvacritxfl_util(
  dataset,
  critx_text,
  critxfl_cond,
  counter,
  bcva_range = NULL,
  bcva_uplim = NULL,
  bcva_lowlim = NULL
)

Arguments

dataset

Input dataset (ADBCVA).

critx_text

String containing the text for CRITx variable.

critxfl_cond

String containing R code detailing the criterion to be satisfied for CRITxFL variable to be equal to "Y".

counter

Integer detailing the value of x to use in "CRITxFL".

bcva_range

Numeric vector of length two detailing lower and upper change in BCVA limits (bcva_range will be called in critxfl_cond if the criterion stipulates that change in BCVA lie inside some range).

bcva_uplim

Numeric value detailing highest change in BCVA limit (bcva_uplim will be called in critxfl_cond if the criterion stipulates that change in BCVA lie below some upper limit).

bcva_lowlim

Numeric value detailing lowest change in BCVA limit (bcva_lowlim will be called in critxfl_cond if the criterion stipulates that change in BCVA lie above some lower limit).

Value

The input ADBCVA dataset with additional columns CRITx, CRITxFL.

Details

The criterion for change in BCVA in CRITxFL can be of three types: (1) value lies within some range; a <= CHG <= b; (2) value is below some upper limit; CHG <= a; (3) value is above some lower limit; b <= CHG. For (1), bcva_range must be specified to this function; for (2), bcva_uplim; for (3) bcva_lowlim. It is only ever necessary to supply one of these three arguments. NOTE: if CHG is equal to NA, then the resulting criterion flag is also marked as NA.

Author

Edoardo Mancini